pyproject.toml
[project]
name = "casino"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]

[tool.uv]
default-groups = ["backend", "fronted", "dev"]


[dependency-groups]
backend = [
    "sanic[ext]>=25.3.0",
    "msgspec>=0.19.0",
    "aiosqlite>=0.21.0",
    "pyjwt>=2.10.1",
]
fronted = []
dev = [
    "nox>=2025.2.9",
    "pyright>=1.1.400",
]
ruff =["ruff>=0.11.7"]
